support exporting VideoClips with a size above 2gb. close #765

+++ AssetStudio/Classes/Mesh.cs
@@ -681,7 +681,7 @@ namespace AssetStudio
             {
                 if (m_VertexData.m_VertexCount > 0)
                 {
-                    var resourceReader = new ResourceReader(m_StreamData.path, assetsFile, m_StreamData.offset, (int)m_StreamData.size);
+                    var resourceReader = new ResourceReader(m_StreamData.path, assetsFile, m_StreamData.offset, m_StreamData.size);
                     m_VertexData.m_DataSize = resourceReader.GetData();
                 }
             }
